Implement all answered QUESTIONS.md items: ILogger, bug fixes, doc cleanup, type fixes
Agent-Logs-Url: https://github.com/archanox/GUNRPG/sessions/9c102004-5b21-46b8-94e9-e8fdc4ac6bbd

+++ GUNRPG.Api/Dtos/ApiProcessOutcomeRequest.cs
@@ -4,7 +4,7 @@ public sealed class ApiProcessOutcomeRequest
 {
     public Guid SessionId { get; init; }
     public bool OperatorDied { get; init; }
-    public int XpGained { get; init; }
+    public long XpGained { get; init; }
     public bool IsVictory { get; init; }
     public List<string> GearLost { get; init; } = new();
 }
